% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_the_future.R
\name{predict_the_future}
\alias{predict_the_future}
\title{predict_the_future}
\usage{
predict_the_future(f, data, start_at_nth = 1, ...)
}
\arguments{
\item{f}{a function that takes a row-wise subset of \code{data} and returns
a scalar prediction}

\item{data}{a data.frame used by \code{f} to make predictions.  Must contain
a column named \code{t} which indicates the time of each observation.}

\item{start_at_nth}{start running the predictions on the \code{n}th
observation. Useful when predictions on fewer than \code{n} observations
are known to be unreliable.}

\item{...}{additional parameters to \code{f}}
}
\value{
a vector of predictions
}
\description{
predict_the_future
}
\examples{
predict_the_future(f = function(t, x) base::mean(x),
                   data = data.frame(t = 1:6, x = 1.3^(1:6)))

}
